refactor: 패키지 구조 수정 (team.guin.api.example => team.guin.example)
