+++ api/src/main/kotlin/team/guin/ApiApplication.kt
@@ -1,11 +1,9 @@
-package team.guin.api
+package team.guin
 
 import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.autoconfigure.domain.EntityScan
 import org.springframework.boot.runApplication
 
 @SpringBootApplication
-@EntityScan(basePackages = ["team.guin.domain.example"])
 class ApiApplication
 
 fun main(args: Array<String>) {
