--- api/src/main/kotlin/team/guin/api/example/controller/request/DeleteRequest.kt
@@ -1,3 +1,0 @@
-package team.guin.api.example.controller.request
-
-data class DeleteRequest(val id: Long)
