--- api/src/main/kotlin/team/guin/api/example/controller/response/AddResponse.kt
@@ -1,5 +1,0 @@
-package team.guin.api.example.controller.response
-
-import team.guin.domain.example.ExampleEntity
-
-data class AddResponse(val added: ExampleEntity)
