--- api/src/main/kotlin/team/guin/api/example/controller/response/ListResponse.kt
@@ -1,5 +1,0 @@
-package team.guin.api.example.controller.response
-
-import team.guin.domain.example.ExampleEntity
-
-data class ListResponse(val list: List<ExampleEntity>)
