+++ api/src/main/kotlin/team/guin/example/controller/ExampleController.kt
@@ -1,4 +1,4 @@
-package team.guin.api.example.controller
+package team.guin.example.controller
 
 import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.GetMapping
@@ -6,8 +6,8 @@ import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
-import team.guin.api.example.controller.request.DeleteRequest
-import team.guin.api.example.service.ExampleService
+import team.guin.example.controller.request.DeleteRequest
+import team.guin.example.service.ExampleService
 
 @RestController
 @RequestMapping("/example")
