+++ api/src/main/kotlin/team/guin/example/controller/request/DeleteRequest.kt
@@ -1,0 +1,3 @@
+package team.guin.example.controller.request
+
+data class DeleteRequest(val id: Long)
