+++ api/src/main/kotlin/team/guin/example/controller/response/AddResponse.kt
@@ -1,0 +1,5 @@
+package team.guin.example.controller.response
+
+import team.guin.example.ExampleEntity
+
+data class AddResponse(val added: ExampleEntity)
