+++ api/src/main/kotlin/team/guin/example/controller/response/ListResponse.kt
@@ -1,0 +1,5 @@
+package team.guin.example.controller.response
+
+import team.guin.example.ExampleEntity
+
+data class ListResponse(val list: List<ExampleEntity>)
