+++ api/src/main/kotlin/team/guin/example/repository/ExampleRepository.kt
@@ -1,8 +1,8 @@
-package team.guin.api.example.repository
+package team.guin.example.repository
 
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
-import team.guin.domain.example.ExampleEntity
+import team.guin.example.ExampleEntity
 
 @Repository
 interface ExampleRepository : JpaRepository<ExampleEntity, Long>
