+++ api/src/test/kotlin/team/guin/ApiApplicationTests.kt
@@ -1,4 +1,4 @@
-package team.guin.api
+package team.guin
 
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
