+++ api/src/test/kotlin/team/guin/example/service/ExampleServiceTests.kt
@@ -1,9 +1,10 @@
-package team.guin.api.example.service
+package team.guin.example.service
 
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.shouldNotBe
 import org.springframework.boot.test.context.SpringBootTest
+import team.guin.example.service.ExampleService
 
 @SpringBootTest
 class ExampleServiceTests(
