+++ domain/src/main/kotlin/team/guin/example/ExampleEntity.kt
@@ -1,4 +1,4 @@
-package team.guin.domain.example
+package team.guin.example
 
 import javax.persistence.Entity
 import javax.persistence.GeneratedValue
